Spell parser refactoring - intremediate step
