+++ src/act.wizard.cpp
@@ -79,6 +79,7 @@
 #include "zone.table.hpp"
 #include "classes/constants.hpp"
 #include "spells.info.h"
+#include "magic.rooms.hpp"
 
 #include <boost/format.hpp>
 #include <boost/algorithm/string.hpp>
@@ -946,21 +947,9 @@ void is_empty_ch(zone_rnum zone_nr, CHAR_DATA *ch)
 	int rnum_start, rnum_stop;
 	bool found = false;
 	CHAR_DATA *caster;
-	//Проверим, нет ли в зоне метки для врат, чтоб не абузили.
-	for (auto it = RoomSpells::aff_room_list.begin(); it != RoomSpells::aff_room_list.end(); ++it)
-	{
-		const auto aff = find_room_affect(*it, SPELL_RUNE_LABEL);
-		if (((*it)->zone == zone_nr)
-			&& aff != (*it)->affected.end())
-		{
-			// если в зоне метка
-			caster = find_char((*aff)->caster_id);
-			if (caster)
-			{
-				sprintf(buf2, "В зоне vnum:%d клетка vnum: %d находится рунная метка игрока: %s.\r\n", zone_table[zone_nr].number, (*it)->number, GET_NAME(caster));
-				send_to_char(buf2, ch);
-			}
-		}
+
+	if (RoomSpells::isZoneRoomAffected(zone_nr, SPELL_RUNE_LABEL)) {
+		send_to_char("В зоне имеется рунная метка.\r\n", ch);
 	}
 
 	for (i = descriptor_list; i; i = i->next)
