+++ src/dg_script/dg_mobcmd.cpp
@@ -40,7 +40,7 @@
 #include "object.prototypes.hpp"
 #include "room.hpp"
 #include "skills.h"
-#include "spell_parser.hpp"
+#include "magic.utils.hpp"
 #include "spells.h"
 #include "structs.h"
 #include "sysdep.h"
