+++ src/interpreter.cpp
@@ -35,6 +35,7 @@
 #include "cmd/create.h"
 #include "cmd/mixture.h"
 #include "cmd/cast.h"
+#include "cmd/employ.h"
 #include "comm.h"
 #include "constants.h"
 #include "craft/craft.commands.hpp"
@@ -328,7 +329,7 @@ void do_toggle(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_sense(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_unban(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_ungroup(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
-void do_use(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
+void do_employ(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_users(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_visible(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
 void do_vnum(CHAR_DATA *ch, char *argument, int cmd, int subcmd);
@@ -572,7 +573,7 @@ cpp_extern const struct command_info cmd_info[] =
 	{"затоптать", POS_STANDING, do_extinguish, 0, 0, 0},
 	{"заточить", POS_RESTING, do_upgrade, 0, 0, 500},
 	{"заучить", POS_RESTING, do_memorize, 0, 0, 0},
-	{"зачитать", POS_RESTING, do_use, 0, SCMD_RECITE, 500},
+	{"зачитать", POS_RESTING, do_employ, 0, SCMD_RECITE, 500},
 	{"зачаровать", POS_STANDING, do_spell_capable, 1, 0, 0},
 	{"зачистить", POS_DEAD, do_sanitize, LVL_GRGOD, 0, 0},
 	{"золото", POS_RESTING, do_gold, 0, 0, 0},
@@ -584,7 +585,7 @@ cpp_extern const struct command_info cmd_info[] =
 	{"изгнать нежить", POS_RESTING, do_turn_undead, 0, 0, -1},
 	{"изучить", POS_SITTING, do_learn, 0, 0, 0},
 	{"информация", POS_SLEEPING, do_gen_ps, 0, SCMD_INFO, 0},
-	{"испить", POS_RESTING, do_use, 0, SCMD_QUAFF, 500},
+	{"испить", POS_RESTING, do_employ, 0, SCMD_QUAFF, 500},
 	{"использовать", POS_RESTING, do_style, 0, 0, 0},
 	{"имя", POS_SLEEPING, do_name, LVL_IMMORT, 0, 0},
 
@@ -650,7 +651,7 @@ cpp_extern const struct command_info cmd_info[] =
 	{"осмотреть", POS_RESTING, do_examine, 0, 0, 0},
 	{"оседлать", POS_STANDING, do_horsetake, 1, 0, -1},
 	{"оскорбить", POS_RESTING, do_insult, 0, 0, -1},
-	{"осушить", POS_RESTING, do_use, 0, SCMD_QUAFF, 300},
+	{"осушить", POS_RESTING, do_employ, 0, SCMD_QUAFF, 300},
 	{"освежевать", POS_STANDING, do_makefood,             0, 0,                      -1},
 	{"ответить", POS_RESTING,    do_reply,                0, 0,                      -1},
 	{"отразить", POS_FIGHTING,   do_multyparry,           0, 0,                      -1},
@@ -714,7 +715,7 @@ cpp_extern const struct command_info cmd_info[] =
 	{"привязать", POS_RESTING, do_horseput, 0, 0, 500},
 	{"приглядеться", POS_RESTING, do_looking, 0, 0, 250},
 	{"прикрыть", POS_FIGHTING, do_protect, 0, 0, -1},
-	{"применить", POS_SITTING, do_use, 1, SCMD_USE, 400},
+	{"применить", POS_SITTING, do_employ, 1, SCMD_USE, 400},
 	//{"прикоснуться", POS_STANDING, do_touch_stigma, 0, 0, -1},
 	{"присесть", POS_RESTING, do_sit, 0, 0, -1},
 	{"прислушаться", POS_RESTING, do_hearing, 0, 0, 300},
@@ -953,13 +954,13 @@ cpp_extern const struct command_info cmd_info[] =
 	{"purge", POS_DEAD, do_purge, LVL_GOD, 0, 0},
 	{"put", POS_RESTING, do_put, 0, 0, 500},
 //	{"python", POS_DEAD, do_console, LVL_GOD, 0, 0},
-	{"quaff", POS_RESTING, do_use, 0, SCMD_QUAFF, 500},
+	{"quaff", POS_RESTING, do_employ, 0, SCMD_QUAFF, 500},
 	{"qui", POS_SLEEPING, do_quit, 0, 0, 0},
 	{"quit", POS_SLEEPING, do_quit, 0, SCMD_QUIT, -1},
 	{"read", POS_RESTING, do_look, 0, SCMD_READ, 200},
 	{"receive", POS_STANDING, do_not_here, 1, 0, -1},
 	{"recipes", POS_RESTING, do_recipes, 0, 0, 0},
-	{"recite", POS_RESTING, do_use, 0, SCMD_RECITE, 500},
+	{"recite", POS_RESTING, do_employ, 0, SCMD_RECITE, 500},
 	{"redit", POS_DEAD, do_olc, 0, SCMD_OLC_REDIT, 0},
 	{"register", POS_DEAD, do_wizutil, LVL_IMMORT, SCMD_REGISTER, 0},
 	{"unregister", POS_DEAD, do_wizutil, LVL_IMMORT, SCMD_UNREGISTER, 0},
@@ -1032,7 +1033,7 @@ cpp_extern const struct command_info cmd_info[] =
 	{"ungroup", POS_DEAD, do_ungroup, 0, 0, -1},
 	{"unlock", POS_SITTING, do_gen_door, 0, SCMD_UNLOCK, 500},
 	{"uptime", POS_DEAD, do_date, LVL_IMMORT, SCMD_UPTIME, 0},
-	{"use", POS_SITTING, do_use, 1, SCMD_USE, 500},
+	{"use", POS_SITTING, do_employ, 1, SCMD_USE, 500},
 	{"users", POS_DEAD, do_users, LVL_IMMORT, 0, 0},
 	{"value", POS_STANDING, do_not_here, 0, 0, -1},
 	{"version", POS_DEAD, do_gen_ps, 0, SCMD_VERSION, 0},
