+++ src/magic.items.h
@@ -1,0 +1,9 @@
+#ifndef __MAGIC_ITEMS_H__
+#define __MAGIC_ITEMS_H__
+
+class CHAR_DATA;
+class OBJ_DATA;
+
+void employMagicItem(CHAR_DATA * ch, OBJ_DATA * obj, const char *argument);
+
+#endif //BYLINS_CREATE_H
