+++ src/modify.cpp
@@ -18,7 +18,7 @@
 #include "handler.h"
 #include "db.h"
 #include "comm.h"
-#include "spell_parser.hpp"
+#include "magic.utils.hpp"
 #include "spells.h"
 #include "mail.h"
 #include "boards/boards.h"
@@ -53,7 +53,7 @@ void show_string(DESCRIPTOR_DATA * d, char *input);
 #define PARSE_EDIT        7
 
 extern const char *MENU;
-extern const char *unused_spellname;	// spell_parser.cpp
+extern const char *unused_spellname;
 
 // local functions
 void smash_tilde(char *str);
