+++ src/morph.cpp
@@ -5,7 +5,7 @@
 #include "structs.h"
 #include "interpreter.h"
 #include "handler.h"
-#include "spell_parser.hpp"
+#include "magic.utils.hpp"
 #include "spells.h"
 #include "chars/char.hpp"
 #include "comm.h"
