+++ src/spec_procs.cpp
@@ -37,7 +37,7 @@
 #include "obj.hpp"
 #include "screen.h"
 #include "skills.h"
-#include "spell_parser.hpp"
+#include "magic.utils.hpp"
 #include "spells.h"
 #include "structs.h"
 #include "sysdep.h"
